feat(engine): engine-level UI scene renderer in psx-engine::ui
Promote the cooked-UI-node renderer out of editor-playtest/overlay.rs
into psx-engine::ui::draw_scene, which renders any &[LevelUiNodeRecord]
slice. Decoupled from the example asset table via two caller resolvers:
a texture lookup (AssetId to UiTextureSlot) and a value-binding resolver.
overlay.rs delegates and keeps its legacy fallback HUD. psx-font becomes
a psx-engine dependency.
